==> 99-update-super-srag-25-05-to-28-11-2021/01-getting-and-writing-in-gcs-auxiliar-datasets-UPDATE.ipynb <==
### imports

import pandas as pd
import pyspark.sql.functions as F
from pyspark.sql.types import DateType
#from pyspark.sql.functions import min, max

pd.set_option('display.float_format', lambda x: '%.2f' % x)

spark = SparkSession.builder.getOrCreate()
TMP_BUCKET = "ai-covid-tmp"
spark.conf.set("temporaryGcsBucket", TMP_BUCKET)
spark.sparkContext.setCheckpointDir("hdfs:///tmp/")
sc = spark.sparkContext

### download and read

!hdfs dfs -mkdir -p /googlemobility/data/
!wget 'https://www.gstatic.com/covid19/mobility/Global_Mobility_Report.csv'
!mv -fv 'Global_Mobility_Report.csv' 'Global_Mobility_Report-28-11-2021.csv'
!hdfs dfs -put 'Global_Mobility_Report-28-11-2021.csv'  /googlemobility/data
df = spark.read.csv("hdfs:///googlemobility/data/Global_Mobility_Report-28-11-2021.csv", header=True, inferSchema= True)

# puting ibge codes on municipalities and states
# file from ftp://geoftp.ibge.gov.br/organizacao_do_territorio/estrutura_territorial/divisao_territorial/2018/DTB_2018.zip
!hdfs dfs -put /RELATORIO_DTB_BRASIL_MUNICIPIO.csv /googlemobility/data/
ibge = spark.read.csv('hdfs:///googlemobility/data/RELATORIO_DTB_BRASIL_MUNICIPIO.csv', header=True, inferSchema= True)
ibge = ibge.withColumnRenamed('Nome_Município', 'mun_name')
ibge = ibge.withColumnRenamed('Código Município Completo', 'mun_code')

### writing on Google Cloud Storage

df.write.csv('gs://ai-covid19-datalake/raw/google-mobility/google-mobility_report-28-11-2021.csv', header=True, mode='overwrite')

ibge.write.csv('gs://ai-covid19-datalake/raw/ibge-data/ibge-municipality-to-code-28-11-2021.csv', header=True, mode='overwrite')

print(ibge.count())
print(ibge.dropDuplicates(['Nome_UF','mun_name']).count())

# FAZER JOIN USANDO O NOME DA CIDADE!
# cond = 
df_joined = df.join(ibge, df['sub_region_2']==ibge['mun_name'], 'left')

df.limit(2).toPandas()

df_joined.filter(F.col('country_region') == 'Brazil').limit(2).toPandas()

df_joined.select('country_region').groupBy('country_region').count().show()

### some investigations

print(df.count())

# Is there Brazil's data? 
df = df.filter(F.col('country_region_code') == "BR")
print(df.count())
# 904413

# which period? 
df = df.withColumn('date', df.date.cast(DateType()))

# df.groupBy('date').count().orderBy('date').show()
maxdate, mindate = df.select(F.max("date"), F.min("date")).first()

print(mindate, maxdate)
# 2020-02-15 2021-05-24

df.printSchema()

# how much states? 

estados = df.select('sub_region_1').distinct()

print(estados.count())
estados.show(28)

cidades = df.select('sub_region_2').distinct()

# how much cities? 

cidades = df.select('sub_region_2').distinct()

print(cidades.count())
cidades.show(20)
# 2254

# from which states, mostly?
