==> 99-update-super-srag-25-05-to-28-11-2021/02-preprocessing-auxiliar-datasets-UPDATE.ipynb <==
import pandas as pd
import pyspark.sql.functions as F
from pyspark.sql import Window
import re
from pyspark.sql.types import *
import unicodedata
from google.cloud import storage
from functools import reduce
from pyspark.sql import DataFrame
import datetime
# spark.conf.set("spark.sql.crossJoin.enabled", True)

<hr />

### google mobility preprocessing

<hr />

# read data from gcs
gmr = spark.read.csv('gs://ai-covid19-datalake/raw/google-mobility/google-mobility_report-28-11-2021.csv', inferSchema=True, header=True)

# head? 

gmr.limit(5).toPandas()

# schema
gmr.printSchema()

# change timestamp var type to date type (we do not need hours or minutes)
gmr = gmr.withColumn('date', F.col('date').cast('date'))

# filtering observations after 24-05-2021
gmr = gmr.filter(F.col('date') > '2021-05-24')

gmr.printSchema()

#### Next steps: 
* select brazillian observations, keeping the header
* make a rapid descriptive analysis
* take out special characters
* upper case the text values
* take the sufix from states names
* write the preprocessed data in a proper location
* make another descriptive analysis

<hr />

# select brazillian observations, keeping the header
print(gmr.count())
gmr = gmr.filter(F.col('country_region_code')=='BR')
print(gmr.count())
# Version until 24-05-2021
# 5561418
# 913701

# Updated version without Filter
# 7796287
# 1301098

<hr />

#### make a rapid descriptive analysis 
* what are the states? 
* how much cities? 
* how much cities from each states? 
* how much observavions from each state? 
* max and min period for observations of each state
* null values per state on variables: census_fips_code, residential_percent_change_from_baseline, workplaces_percent_change_from_baseline, transit_stations_percent_change_from_baseline, parks_percent_change_from_baseline, and retail_and_recreation_percent_change_from_baseline

# what are the states?
# how much cities?
print("how much cities:", gmr.select(['sub_region_2']).distinct().count())

# # how much cities from each states?
states_and_cities = gmr.select(['sub_region_1', 'sub_region_2']).distinct().groupby('sub_region_1').count().sort('sub_region_1').withColumnRenamed('count', 'n_cities')
# states_and_cities.show(30)

# # how much observavions from each state?
n_observations = gmr.select(['sub_region_1', 'sub_region_2']).groupby('sub_region_1').count().sort('sub_region_1').withColumnRenamed('count', 'n_observations')
# n_observations.show(30)

# # max and min period for observations of each state
w = Window.partitionBy('sub_region_1')
min_date = gmr.select(['sub_region_1', 'date']).withColumn('min_date', F.min('date').over(w)).where(F.col('date') == F.col('min_date')).distinct().drop('date')
max_date = gmr.select(['sub_region_1', 'date']).withColumn('max_date', F.max('date').over(w)).where(F.col('date') == F.col('max_date')).distinct().drop('date')
# min_date.show(30)
# max_date.show(30)

# null values
columns = ['census_fips_code', 'residential_percent_change_from_baseline', 'workplaces_percent_change_from_baseline', 'transit_stations_percent_change_from_baseline', 'parks_percent_change_from_baseline', 'retail_and_recreation_percent_change_from_baseline']

nulls_census_fips_code = gmr.filter(F.col('census_fips_code').isNull()).select(['sub_region_1', 'census_fips_code']).groupBy(['sub_region_1']).count().withColumnRenamed('count', 'nulls_census_fips_code')
nulls_residential_percent_change_from_baseline = gmr.filter(F.col('residential_percent_change_from_baseline').isNull()).select(['sub_region_1', 'residential_percent_change_from_baseline']).groupBy(['sub_region_1']).count().withColumnRenamed('count', 'null_residential_percent_change_from_baseline')
nulls_workplaces_percent_change_from_baseline = gmr.filter(F.col('workplaces_percent_change_from_baseline').isNull()).select(['sub_region_1', 'workplaces_percent_change_from_baseline']).groupBy(['sub_region_1']).count().withColumnRenamed('count', 'workplaces_percent_change_from_baseline')
nulls_transit_stations_percent_change_from_baseline = gmr.filter(F.col('transit_stations_percent_change_from_baseline').isNull()).select(['sub_region_1', 'transit_stations_percent_change_from_baseline']).groupBy(['sub_region_1']).count().withColumnRenamed('count', 'transit_stations_percent_change_from_baseline')
nulls_parks_percent_change_from_baseline = gmr.filter(F.col('parks_percent_change_from_baseline').isNull()).select(['sub_region_1', 'parks_percent_change_from_baseline']).groupBy(['sub_region_1']).count().withColumnRenamed('count', 'parks_percent_change_from_baseline')
nulls_retail_and_recreation_percent_change_from_baseline = gmr.filter(F.col('retail_and_recreation_percent_change_from_baseline').isNull()).select(['sub_region_1', 'retail_and_recreation_percent_change_from_baseline']).groupBy(['sub_region_1']).count().withColumnRenamed('count', 'retail_and_recreation_percent_change_from_baseline')

nulls = nulls_census_fips_code.join(nulls_residential_percent_change_from_baseline, 'sub_region_1', how='left')\
                                .join(nulls_workplaces_percent_change_from_baseline, 'sub_region_1', how='left')\
                                .join(nulls_transit_stations_percent_change_from_baseline, 'sub_region_1', how='left')\
                                .join(nulls_parks_percent_change_from_baseline, 'sub_region_1', how='left')\
                                .join(nulls_retail_and_recreation_percent_change_from_baseline, 'sub_region_1', how='left')
                                

# joining all
desc_gmr = states_and_cities.join(n_observations, 'sub_region_1', 'left')\
                            .join(min_date, 'sub_region_1', 'left')\
                            .join(max_date, 'sub_region_1', 'left')\
                            .join(nulls, 'sub_region_1', 'left')

# writing and showing

desc_gmr.write.csv('gs://ai-covid19-datalake/raw/google-mobility/desc_pp_google_mobility_report_28_11_2021_pre_formatting_v1.csv', header=True, mode='overwrite')

desc_gmr.toPandas()

<hr />

# take out special characters
def remove_accents(col):
    col = str(col)
    col = col.replace('á', 'a')
    col = col.replace('à', 'a')
    col = col.replace('ã', 'a')
    col = col.replace('â', 'a')
    col = col.replace('é', 'e')
    col = col.replace('ê', 'e')
    col = col.replace('í', 'i')
    col = col.replace('ó', 'o')
    col = col.replace('õ', 'o')
    col = col.replace('ô', 'o')
    col = col.replace('ú', 'u')
    col = col.replace('ü', 'u')
    col = col.replace('ç', 'c')
    return col
udf_remove_accents = F.udf(remove_accents, StringType())

def normalize_ascii(col):
    return re.sub('[^A-Za-z0-9]+', ' ', str(col))
udf_normalize_ascii = F.udf(normalize_ascii, StringType())

def take_state_prefix(col):
    return col.replace('STATE OF ', '')
udf_take_state_prefix = F.udf(take_state_prefix, StringType())

<hr />

# dropping NA in state and municipality
gmr = gmr.dropna(subset=['sub_region_1', 'sub_region_2'])

# taking accents and non-ascii chars
gmr = gmr.withColumn('sub_region_1', udf_remove_accents(F.col('sub_region_1')))
gmr = gmr.withColumn('sub_region_1', udf_normalize_ascii(F.col('sub_region_1')))

gmr = gmr.withColumn('sub_region_2', udf_remove_accents(F.col('sub_region_2')))
gmr = gmr.withColumn('sub_region_2', udf_normalize_ascii(F.col('sub_region_2')))


# turning columns into upper case
gmr = gmr.withColumn('sub_region_1', F.upper(F.col('sub_region_1')))
gmr = gmr.withColumn('sub_region_2', F.upper(F.col('sub_region_2')))

# taking prefix
gmr = gmr.withColumn('sub_region_1', udf_take_state_prefix(F.col('sub_region_1')))

gmr.select(['sub_region_1', 'sub_region_2']).distinct().show()

# what are the states?
# how much cities?
print("how much cities:", gmr.select(['sub_region_2']).distinct().count())

# # how much cities from each states?
states_and_cities = gmr.select(['sub_region_1', 'sub_region_2']).distinct().groupby('sub_region_1').count().sort('sub_region_1').withColumnRenamed('count', 'n_cities')
# states_and_cities.show(30)

# # how much observavions from each state?
n_observations = gmr.select(['sub_region_1', 'sub_region_2']).groupby('sub_region_1').count().sort('sub_region_1').withColumnRenamed('count', 'n_observations')
# n_observations.show(30)

# # max and min period for observations of each state
w = Window.partitionBy('sub_region_1')
min_date = gmr.select(['sub_region_1', 'date']).withColumn('min_date', F.min('date').over(w)).where(F.col('date') == F.col('min_date')).distinct().drop('date')
max_date = gmr.select(['sub_region_1', 'date']).withColumn('max_date', F.max('date').over(w)).where(F.col('date') == F.col('max_date')).distinct().drop('date')
# min_date.show(30)
# max_date.show(30)

# null values
columns = ['census_fips_code', 'residential_percent_change_from_baseline', 'workplaces_percent_change_from_baseline', 'transit_stations_percent_change_from_baseline', 'parks_percent_change_from_baseline', 'retail_and_recreation_percent_change_from_baseline']

nulls_census_fips_code = gmr.filter(F.col('census_fips_code').isNull()).select(['sub_region_1', 'census_fips_code']).groupBy(['sub_region_1']).count().withColumnRenamed('count', 'nulls_census_fips_code')
nulls_residential_percent_change_from_baseline = gmr.filter(F.col('residential_percent_change_from_baseline').isNull()).select(['sub_region_1', 'residential_percent_change_from_baseline']).groupBy(['sub_region_1']).count().withColumnRenamed('count', 'null_residential_percent_change_from_baseline')
nulls_workplaces_percent_change_from_baseline = gmr.filter(F.col('workplaces_percent_change_from_baseline').isNull()).select(['sub_region_1', 'workplaces_percent_change_from_baseline']).groupBy(['sub_region_1']).count().withColumnRenamed('count', 'workplaces_percent_change_from_baseline')
nulls_transit_stations_percent_change_from_baseline = gmr.filter(F.col('transit_stations_percent_change_from_baseline').isNull()).select(['sub_region_1', 'transit_stations_percent_change_from_baseline']).groupBy(['sub_region_1']).count().withColumnRenamed('count', 'transit_stations_percent_change_from_baseline')
nulls_parks_percent_change_from_baseline = gmr.filter(F.col('parks_percent_change_from_baseline').isNull()).select(['sub_region_1', 'parks_percent_change_from_baseline']).groupBy(['sub_region_1']).count().withColumnRenamed('count', 'parks_percent_change_from_baseline')
nulls_retail_and_recreation_percent_change_from_baseline = gmr.filter(F.col('retail_and_recreation_percent_change_from_baseline').isNull()).select(['sub_region_1', 'retail_and_recreation_percent_change_from_baseline']).groupBy(['sub_region_1']).count().withColumnRenamed('count', 'retail_and_recreation_percent_change_from_baseline')

nulls = nulls_census_fips_code.join(nulls_residential_percent_change_from_baseline, 'sub_region_1', how='left')\
                                .join(nulls_workplaces_percent_change_from_baseline, 'sub_region_1', how='left')\
                                .join(nulls_transit_stations_percent_change_from_baseline, 'sub_region_1', how='left')\
                                .join(nulls_parks_percent_change_from_baseline, 'sub_region_1', how='left')\
                                .join(nulls_retail_and_recreation_percent_change_from_baseline, 'sub_region_1', how='left')
                                

# joining all
desc_gmr = states_and_cities.join(n_observations, 'sub_region_1', 'left')\
                            .join(min_date, 'sub_region_1', 'left')\
                            .join(max_date, 'sub_region_1', 'left')\
                            .join(nulls, 'sub_region_1', 'left')

# writing and showing

desc_gmr.write.csv('gs://ai-covid19-datalake/standard/google-mobility/desc_pp_google_mobility_report_28_11_2021_pos_formatting_v1.csv', header=True, mode='overwrite')

desc_gmr.toPandas()

gmr.write.csv('gs://ai-covid19-datalake/standard/google-mobility/pp_google-mobility_report-28-11-2021.csv', header=True, mode='overwrite')

<hr />

### municipality ibge data

<hr />

# read data from gcs
ibge = spark.read.csv('gs://ai-covid19-datalake/raw/ibge-data/ibge-municipality-to-code-28-11-2021.csv', inferSchema=True, header=True)

# show some lines
ibge.limit(5).toPandas()

# schema
ibge.printSchema()

ibge = ibge.select(['UF', 'mun_code', 'Nome_UF', 'mun_name']).withColumnRenamed('Nome_UF', 'sub_region_1').withColumnRenamed('mun_name', 'sub_region_2')

# dropping NA in state and municipality
ibge = ibge.dropna(subset=['sub_region_1', 'sub_region_2'])

# taking accents and non-ascii chars
ibge = ibge.withColumn('sub_region_1', udf_remove_accents(F.col('sub_region_1')))
ibge = ibge.withColumn('sub_region_1', udf_normalize_ascii(F.col('sub_region_1')))

ibge = ibge.withColumn('sub_region_2', udf_remove_accents(F.col('sub_region_2')))
ibge = ibge.withColumn('sub_region_2', udf_normalize_ascii(F.col('sub_region_2')))


# turning columns into upper case
ibge = ibge.withColumn('sub_region_1', F.upper(F.col('sub_region_1')))
ibge = ibge.withColumn('sub_region_2', F.upper(F.col('sub_region_2')))

# show some lines
ibge.limit(5).toPandas()

ibge.write.csv('gs://ai-covid19-datalake/standard/ibge-data/pp_ibge-municipality-to-code-28-11-2021.csv', header=True, mode='overwrite')

<hr />

### meteorological data

<hr />

# getting the list of inmet files

client = storage.Client()
BUCKET_NAME = 'ai-covid19-datalake'
bucket = client.get_bucket(BUCKET_NAME)

blobs = bucket.list_blobs()
files = []
for blob in blobs:
    uri = 'raw/inmet-data/inmet-meteorological-data-25052021-a-28112021/'
    if blob.name[0:-38] == uri:
        files.append('gs://ai-covid19-datalake/'+blob.name)

inmet_schema = StructType([
    StructField('measurement_date', DateType()),
    StructField('total_daily_precipitation_mm', FloatType()),
    StructField('daily_atmospheric_pression_mb', FloatType()),
    StructField('daily_avg_dew_point_temp_c', FloatType()),
    StructField('max_daily_temp_maxima_diaria_c', FloatType()),
    StructField('daily_avg_temp_c', FloatType()),
    StructField('daily_min_temp_c', FloatType()),
    StructField('daily_avg_relative_air_humidity_percent', FloatType()),
    StructField('daily_min_relative_air_humidity_percent', FloatType()),
    StructField('max_gust_wind_ms', FloatType()),
    StructField('avg_wind_velocity_ms', FloatType()),
    StructField('empty', StringType())
    
])
    
inmet_header = ['measurement_date', 'total_daily_precipitation_mm', 'daily_atmospheric_pression_mb', 'daily_avg_dew_point_temp_c',
          'max_daily_temp_maxima_diaria_c', 'daily_avg_temp_c', 'daily_min_temp_c', 'daily_avg_relative_air_humidity_percent',
          'daily_min_relative_air_humidity_percent', 'max_gust_wind_ms', 'avg_wind_velocity_ms']

def replace_comma_and_convert(line):
    positions = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10]
    date_positions = [0]
    
    for dp in date_positions:
        # converting string do proper date type, e.g.: '2019-01-02'
        if (line[dp] != 'null') and (line[dp] != ''):
            date_time_obj = datetime.datetime.strptime(line[dp], '%Y-%m-%d')
            line[dp] = date_time_obj.date()
        else:
            line[dp] = None

    
    for p in positions:
        if (line[p] != 'null') and (line[p] != ''):
            line[p] = float(line[p].replace(',', '.'))
        else:
            line[p] = None
    return line

dfs = []
for f in files:
    # reading file from list
    inmet_data = sc.textFile(f).map(lambda x: x.split(';'))
    
    # taking out blank lines (serious matter)
    inmet_data = inmet_data.filter(lambda x: x != [''])
    
    # getting the station ids 
    station_name = inmet_data.filter(lambda x: "Nome:" in x[0]).collect()[0][0].split(' ')[1]
    station_code = inmet_data.filter(lambda x: "Estacao:" in x[0]).collect()[0][0].split(' ')[2]
    
    # dropping the first 10 unnecessary top lines on file
    inmet_data = inmet_data.zipWithIndex().filter(lambda tup: tup[1] > 10).map(lambda x:x[0])
    
    # converting string to float values
    inmet_data = inmet_data.map(replace_comma_and_convert)
    
    # converting rdd to a dataframe with schema 
    df = spark.createDataFrame(inmet_data, schema=inmet_schema)
    
    # creating station and municipality variables
    df = df.withColumn('station_id', F.lit(station_code)).withColumn('municipality_name', F.lit(station_name))
    
    # appending new dataframe to be joined
    dfs.append(df)
    
inmet_df = reduce(DataFrame.unionAll, dfs)

# built dataset
inmet_df.limit(5).toPandas()

# reading stations link file
stations_link = spark.read.csv('gs://ai-covid19-datalake/raw/inmet-data/inmet-estacoes-19-02-2018.csv', header=True).select(['station_id', 'sub_region_1'])

# list of stations and its states
stations_link.limit(5).toPandas()

# bringing the state and municipality information
inmet_df = inmet_df.join(stations_link, 'station_id', 'left')

# now with the new state column added
inmet_df.limit(5).toPandas()

inmet_df.select('measurement_date').printSchema()

# df.groupBy('date').count().orderBy('date').show()
maxdate, mindate = df.select(F.max("measurement_date"), F.min("measurement_date")).first()

print(mindate, maxdate)

inmet_df.write.csv('gs://ai-covid19-datalake/standard/inmet-data/pp_inmet_meteorological_data-25052021-a-28112021', header=True, mode='overwrite')
