==> 99-update-super-srag-25-05-to-28-11-2021/04-merging-to-super-srag-UPDATE.ipynb <==
import pandas as pd
import pyspark.sql.functions as F
from pyspark.sql.types import *

pd.set_option("display.max_rows", 101)
pd.set_option("display.max_columns", 101)

<hr />

### reading preprocessed dataframes

srag_2021 = spark.read.parquet('gs://ai-covid19-datalake/standard/srag/pp_interm_srag_2021_v3_new_attr-UPDATE-28-11-2021')

gmr = spark.read.csv('gs://ai-covid19-datalake/standard/google-mobility/pp_google-mobility_report-28-11-2021.csv', header=True, inferSchema=True)

ibge = spark.read.csv('gs://ai-covid19-datalake/standard/ibge-data/pp_ibge-municipality-to-code-28-11-2021.csv', header=True, inferSchema=True)

inmet = spark.read.csv('gs://ai-covid19-datalake/standard/inmet-data/pp_inmet_meteorological_data-25052021-a-28112021', header=True, inferSchema=True)

epi_weeks = spark.read.csv('gs://ai-covid19-datalake/raw/epidemiological-data/epi_weeks.csv', header=True, inferSchema=True)

<hr />

# Preparing srag to receive the new information from UF's and Epidemiological Weeks

# # dropping duplicate columns
# def dropDupeDfCols(df):
#     newcols = []
#     dupcols = []

#     for i in range(len(df.columns)):
#         if df.columns[i] not in newcols:
#             newcols.append(df.columns[i])
#         else:
#             dupcols.append(i)

#     df = df.toDF(*[str(i) for i in range(len(df.columns))])
#     for dupcol in dupcols:
#         df = df.drop(str(dupcol))

#     return df.toDF(*newcols)

# srag_2019 = dropDupeDfCols(srag_2019)
# srag_2020 = dropDupeDfCols(srag_2020)
# srag_2021 = dropDupeDfCols(srag_2021)

srag_2021 = srag_2021.withColumn('ANO', F.lit('2021'))

def get_epi_week_year(epi_week, year):
    return str(epi_week) + '-' + str(year)
udf_get_epi_week_year = F.udf(get_epi_week_year, StringType())

# adding 'epi_week_year' tag
srag_2021 = srag_2021.withColumn('epi_week_year', udf_get_epi_week_year(F.col('SEM_PRI'), F.col('ANO')))

# padding
srag_2021 = srag_2021.withColumn('epi_week_year', F.lpad(F.col('epi_week_year'), 7, '0'))

def get_uf(cod_mun_res):
    return str(cod_mun_res)[:2]
udf_get_uf = F.udf(get_uf, StringType())

srag_2021 = srag_2021.withColumn('UF', udf_get_uf(F.col('CO_MUN_NOT')))

# Google Mobility Report with state codes by epi_weeks

###  adding epidemiological weeks

# converting date type variables
gmr = gmr.withColumn('date', F.to_date(F.col('date'), 'dd/MM/yyyy'))
epi_weeks = epi_weeks.withColumn('date', F.to_date(F.col('date'), 'dd/MM/yyyy'))

# find the date from 7 or 14 days before
epi_weeks = epi_weeks.withColumn('date_7_days_early', F.date_sub(F.col('date'), 7))
epi_weeks = epi_weeks.withColumn('date_14_days_early', F.date_sub(F.col('date'), 14))

# finding the epi_week number from 1 or 2 weeks before
def find_one_or_two_epi_weeks(n_weeks, col):
    return col - n_weeks
udf_find_one_or_two_epi_weeks = F.udf(find_one_or_two_epi_weeks, IntegerType())

epi_weeks = epi_weeks.withColumn('epi_week_7_days_early', udf_find_one_or_two_epi_weeks(F.lit(1), F.col('epi_week')))
epi_weeks = epi_weeks.withColumn('epi_week_14_days_early', udf_find_one_or_two_epi_weeks(F.lit(2), F.col('epi_week')))

# creating the epi_week tag for the new found early epi_weeks
def set_epi_week_year(col1, col2):
    string = str(col1) + '-' + str(col2)
    return string
udf_set_epi_week_year = F.udf(set_epi_week_year, StringType())

epi_weeks = epi_weeks.withColumn('epi_week_year_7_days_early', udf_set_epi_week_year(F.col('epi_week_7_days_early'), F.col('epi_year')))
epi_weeks = epi_weeks.withColumn('epi_week_year_14_early', udf_set_epi_week_year(F.col('epi_week_14_days_early'), F.col('epi_year')))

# make sure to erase inexistent epi_week
epi_weeks = epi_weeks.withColumn('epi_week_year_7_days_early', F.when((F.col('epi_week_7_days_early') < 1), None).otherwise(F.col('epi_week_year_7_days_early')))
epi_weeks = epi_weeks.withColumn('epi_week_7_days_early', F.when((F.col('epi_week_7_days_early') < 1), None).otherwise(F.col('epi_week_7_days_early')))

epi_weeks = epi_weeks.withColumn('epi_week_year_14_early', F.when((F.col('epi_week_14_days_early') < 1), None).otherwise(F.col('epi_week_year_14_early')))
epi_weeks = epi_weeks.withColumn('epi_week_14_days_early', F.when((F.col('epi_week_14_days_early') < 1), None).otherwise(F.col('epi_week_14_days_early')))

# joining using the actual date
epi_weeks_actual = epi_weeks.select(['date', 'epi_week', 'epi_year', 'epi_week_year'])
gmr = gmr.join(epi_weeks_actual, 'date', 'left')

# joining using date from 7 days before
epi_weeks_7 = epi_weeks.select(['date_7_days_early', 'epi_week_7_days_early', 'epi_week_year_7_days_early'])
gmr = gmr.join(epi_weeks_7, F.date_add(gmr.date, 7) == epi_weeks.date_7_days_early, 'left')

# joining using date from 14 days before
epi_weeks_14 = epi_weeks.select(['date_14_days_early', 'epi_week_14_days_early', 'epi_week_year_14_early'])
gmr = gmr.join(epi_weeks_14, F.date_add(gmr.date, 14) == epi_weeks.date_14_days_early, 'left')

epi_weeks.limit(10).toPandas()

gmr.limit(5).toPandas()

### adding state code

gmr = gmr.join(ibge.select('UF', 'sub_region_1'), 'sub_region_1', 'left')

### aggregating

gmr1 = gmr.groupBy(['UF', 'epi_week_year']).agg({"residential_percent_change_from_baseline":"avg",
                                          "workplaces_percent_change_from_baseline":"avg",
                                          "transit_stations_percent_change_from_baseline":"avg",
                                          "parks_percent_change_from_baseline":"avg",
                                          "grocery_and_pharmacy_percent_change_from_baseline":"avg",
                                          "retail_and_recreation_percent_change_from_baseline":"avg"})\
                                    .withColumnRenamed('avg(transit_stations_percent_change_from_baseline)', 'gmr_transit_stations_avg')\
                                    .withColumnRenamed('avg(grocery_and_pharmacy_percent_change_from_baseline)', 'gmr_grocery_and_pharmacy_avg')\
                                    .withColumnRenamed('avg(retail_and_recreation_percent_change_from_baseline)', 'gmr_retail_and_recreation_avg')\
                                    .withColumnRenamed('avg(retail_and_recreation_percent_change_from_baseline)', 'gmr_retail_and_recreation_avg')\
                                    .withColumnRenamed('avg(workplaces_percent_change_from_baseline)', 'gmr_workplaces_percent_avg')\
                                    .withColumnRenamed('avg(residential_percent_change_from_baseline)', 'gmr_residential_percent_avg')\
                                    .withColumnRenamed('avg(parks_percent_change_from_baseline)', 'gmr_parks_percent_avg')\
                                    .orderBy('UF')

gmr7 = gmr.groupBy(['UF', 'epi_week_year_7_days_early']).agg({"residential_percent_change_from_baseline":"avg",
                                          "workplaces_percent_change_from_baseline":"avg",
                                          "transit_stations_percent_change_from_baseline":"avg",
                                          "parks_percent_change_from_baseline":"avg",
                                          "grocery_and_pharmacy_percent_change_from_baseline":"avg",
                                          "retail_and_recreation_percent_change_from_baseline":"avg"})\
                                    .withColumnRenamed('avg(transit_stations_percent_change_from_baseline)', 'gmr_transit_stations_1week_before_avg')\
                                    .withColumnRenamed('avg(grocery_and_pharmacy_percent_change_from_baseline)', 'gmr_grocery_and_pharmacy_1week_before_avg')\
                                    .withColumnRenamed('avg(retail_and_recreation_percent_change_from_baseline)', 'gmr_retail_and_recreation_1week_before_avg')\
                                    .withColumnRenamed('avg(retail_and_recreation_percent_change_from_baseline)', 'gmr_retail_and_recreation_1week_before_avg')\
                                    .withColumnRenamed('avg(workplaces_percent_change_from_baseline)', 'gmr_workplaces_percent_1week_before_avg')\
                                    .withColumnRenamed('avg(residential_percent_change_from_baseline)', 'gmr_residential_percent_1week_before_avg')\
                                    .withColumnRenamed('avg(parks_percent_change_from_baseline)', 'gmr_parks_percent_1week_before_avg')\
                                    .orderBy('UF')

gmr14 = gmr.groupBy(['UF', 'epi_week_year_14_early']).agg({"residential_percent_change_from_baseline":"avg",
                                          "workplaces_percent_change_from_baseline":"avg",
                                          "transit_stations_percent_change_from_baseline":"avg",
                                          "parks_percent_change_from_baseline":"avg",
                                          "grocery_and_pharmacy_percent_change_from_baseline":"avg",
                                          "retail_and_recreation_percent_change_from_baseline":"avg"})\
                                    .withColumnRenamed('avg(transit_stations_percent_change_from_baseline)', 'gmr_transit_stations_2weeks_avg')\
                                    .withColumnRenamed('avg(grocery_and_pharmacy_percent_change_from_baseline)', 'gmr_grocery_and_pharmacy_2weeks_avg')\
                                    .withColumnRenamed('avg(retail_and_recreation_percent_change_from_baseline)', 'gmr_retail_and_recreation_2weeks_avg')\
                                    .withColumnRenamed('avg(retail_and_recreation_percent_change_from_baseline)', 'gmr_retail_and_recreation_2weeks_avg')\
                                    .withColumnRenamed('avg(workplaces_percent_change_from_baseline)', 'gmr_workplaces_percent_2weeks_avg')\
                                    .withColumnRenamed('avg(residential_percent_change_from_baseline)', 'gmr_residential_percent_2weeks_avg')\
                                    .withColumnRenamed('avg(parks_percent_change_from_baseline)', 'gmr_parks_percent_2weeks_avg')\
                                    .orderBy('UF')

##### making some validations

gmr1.filter((F.col('UF')=='11') &  ((F.col('epi_week_year') == '22-2020') | (F.col('epi_week_year') == '23-2020') | (F.col('epi_week_year') == '24-2020') | (F.col('epi_week_year') == '25-2020')))\
    .orderBy('epi_week_year').limit(5).toPandas()

gmr7.filter((F.col('UF')=='11') &  ((F.col('epi_week_year_7_days_early') == '22-2020') | (F.col('epi_week_year_7_days_early') == '23-2020') | (F.col('epi_week_year_7_days_early') == '24-2020') | (F.col('epi_week_year_7_days_early') == '25-2020')))\
    .orderBy('epi_week_year_7_days_early').limit(5).toPandas()

gmr14.filter((F.col('UF')=='11') &  ((F.col('epi_week_year_14_early') == '22-2020') | (F.col('epi_week_year_14_early') == '23-2020') | (F.col('epi_week_year_14_early') == '24-2020') | (F.col('epi_week_year_14_early') == '25-2020')))\
    .orderBy('epi_week_year_14_early').limit(5).toPandas()

# joining all for 1 week early
gmr7 = gmr7.withColumnRenamed('UF', 'UF7')
cond = [gmr1.UF == gmr7.UF7, gmr1.epi_week_year == gmr7.epi_week_year_7_days_early]
gmr_agg = gmr1.join(gmr7, cond, 'left')

gmr_agg.filter((F.col('UF')=='11') &  ((F.col('epi_week_year') == '22-2020') | (F.col('epi_week_year') == '23-2020') | (F.col('epi_week_year') == '24-2020') | (F.col('epi_week_year') == '25-2020')))\
    .orderBy('epi_week_year').limit(5).toPandas()

# joining all for 2 weeks early
gmr14 = gmr14.withColumnRenamed('UF', 'UF14')
cond = [gmr1.UF == gmr14.UF14, gmr1.epi_week_year == gmr14.epi_week_year_14_early]
gmr_agg = gmr_agg.join(gmr14, cond, 'left')

gmr_agg.filter((F.col('UF')=='11') &  ((F.col('epi_week_year') == '22-2020') | (F.col('epi_week_year') == '23-2020') | (F.col('epi_week_year') == '24-2020') | (F.col('epi_week_year') == '25-2020')))\
    .orderBy('epi_week_year').limit(5).toPandas()

# casting to string
gmr_agg = gmr_agg.withColumn('UF', F.col('UF').cast('string'))

# selecting variables of interest
gmr_agg = gmr_agg.select(['UF', 'epi_week_year',
                          'gmr_transit_stations_avg', 'gmr_grocery_and_pharmacy_avg', 'gmr_retail_and_recreation_avg', 'gmr_workplaces_percent_avg', 'gmr_residential_percent_avg', 'gmr_parks_percent_avg',
                          'gmr_transit_stations_1week_before_avg', 'gmr_grocery_and_pharmacy_1week_before_avg', 'gmr_retail_and_recreation_1week_before_avg', 'gmr_workplaces_percent_1week_before_avg', 'gmr_residential_percent_1week_before_avg', 'gmr_parks_percent_1week_before_avg',
                          'gmr_transit_stations_2weeks_avg', 'gmr_grocery_and_pharmacy_2weeks_avg', 'gmr_retail_and_recreation_2weeks_avg', 'gmr_workplaces_percent_2weeks_avg', 'gmr_residential_percent_2weeks_avg', 'gmr_parks_percent_2weeks_avg'])

# padding 'epi_week_year' col
gmr_agg = gmr_agg.withColumn('epi_week_year', F.lpad(F.col('epi_week_year'), 7, '0'))

gmr_agg.limit(10).toPandas()

### Joining mobility data for srags

#### 2021

print("How much distincts before? ", srag_2021.select('NU_NOTIFIC').distinct().count())
print("How much distincts epi_weeks before? ", srag_2021.select('epi_week_year').distinct().count())

srag_2021 = srag_2021.join(gmr_agg, ['UF', 'epi_week_year'], 'left')

print("How much distincts after? ", srag_2021.select('NU_NOTIFIC').distinct().count())
print("How much distincts epi_weeks after? ", srag_2021.select('epi_week_year').distinct().count())

# Meteorological data with state codes by epi_weeks

inmet.limit(5).toPandas()

inmet = inmet.join(ibge.select('UF', 'sub_region_1'), 'sub_region_1', 'left')
inmet = inmet.withColumn('UF', F.col('UF').cast('string'))

inmet = inmet.na.drop(subset=['sub_region_1', 'UF'])

inmet = inmet.withColumnRenamed('measurement_date', 'date')

# joining using the actual date
epi_weeks_actual = epi_weeks.select(['date', 'epi_week', 'epi_year', 'epi_week_year'])
inmet = inmet.join(epi_weeks_actual, 'date', 'left')

# joining using date from 7 days before
epi_weeks_7 = epi_weeks.select(['date_7_days_early', 'epi_week_7_days_early', 'epi_week_year_7_days_early'])
inmet = inmet.join(epi_weeks_7, F.date_add(inmet.date, 7) == epi_weeks.date_7_days_early, 'left')

# joining using date from 14 days before
epi_weeks_14 = epi_weeks.select(['date_14_days_early', 'epi_week_14_days_early', 'epi_week_year_14_early'])
inmet = inmet.join(epi_weeks_14, F.date_add(inmet.date, 14) == epi_weeks.date_14_days_early, 'left')

inmet1 = inmet.groupBy(['UF', 'epi_week_year']).agg({"total_daily_precipitation_mm":"avg",
                                          "daily_avg_temp_c":"avg",
                                          "daily_avg_relative_air_humidity_percent":"avg"})\
                                    .withColumnRenamed('avg(total_daily_precipitation_mm)', 'inmet_daily_precipt_avg')\
                                    .withColumnRenamed('avg(daily_avg_temp_c)', 'inmet_temp_c_avg')\
                                    .withColumnRenamed('avg(daily_avg_relative_air_humidity_percent)', 'inmet_relative_air_humidity_avg')\
                                    .orderBy('UF')

inmet7 = inmet.groupBy(['UF', 'epi_week_year_7_days_early']).agg({"total_daily_precipitation_mm":"avg",
                                          "daily_avg_temp_c":"avg",
                                          "daily_avg_relative_air_humidity_percent":"avg"})\
                                    .withColumnRenamed('avg(total_daily_precipitation_mm)', 'inmet_daily_precipt_1week_before_avg')\
                                    .withColumnRenamed('avg(daily_avg_temp_c)', 'inmet_temp_c_1week_before_avg')\
                                    .withColumnRenamed('avg(daily_avg_relative_air_humidity_percent)', 'inmet_relative_air_humidity_1week_before_avg')\
                                    .orderBy('UF')

inmet14 = inmet.groupBy(['UF', 'epi_week_year_14_early']).agg({"total_daily_precipitation_mm":"avg",
                                          "daily_avg_temp_c":"avg",
                                          "daily_avg_relative_air_humidity_percent":"avg"})\
                                    .withColumnRenamed('avg(total_daily_precipitation_mm)', 'inmet_daily_precipt_2weeks_before_avg')\
                                    .withColumnRenamed('avg(daily_avg_temp_c)', 'inmet_temp_c_2weeks_before_avg')\
                                    .withColumnRenamed('avg(daily_avg_relative_air_humidity_percent)', 'inmet_relative_air_humidity_2weeks_before_avg')\
                                    .orderBy('UF')

##### some validations

inmet1.filter((F.col('UF')=='11') &  ((F.col('epi_week_year') == '22-2020') | (F.col('epi_week_year') == '23-2020') | (F.col('epi_week_year') == '24-2020') | (F.col('epi_week_year') == '25-2020')))\
    .orderBy('epi_week_year').limit(5).toPandas()

inmet7.filter((F.col('UF')=='11') &  ((F.col('epi_week_year_7_days_early') == '22-2020') | (F.col('epi_week_year_7_days_early') == '23-2020') | (F.col('epi_week_year_7_days_early') == '24-2020') | (F.col('epi_week_year_7_days_early') == '25-2020')))\
    .orderBy('epi_week_year_7_days_early').limit(5).toPandas()

inmet14.filter((F.col('UF')=='11') &  ((F.col('epi_week_year_14_early') == '22-2020') | (F.col('epi_week_year_14_early') == '23-2020') | (F.col('epi_week_year_14_early') == '24-2020') | (F.col('epi_week_year_14_early') == '25-2020')))\
    .orderBy('epi_week_year_14_early').limit(5).toPandas()

<hr />

# joining all for 1 week early
inmet7 = inmet7.withColumnRenamed('UF', 'UF7')
cond = [inmet1.UF == inmet7.UF7, inmet1.epi_week_year == inmet7.epi_week_year_7_days_early]
inmet_agg = inmet1.join(inmet7, cond, 'left')

inmet_agg.filter((F.col('UF')=='11') &  ((F.col('epi_week_year') == '22-2020') | (F.col('epi_week_year') == '23-2020') | (F.col('epi_week_year') == '24-2020') | (F.col('epi_week_year') == '25-2020')))\
    .orderBy('epi_week_year').limit(5).toPandas()

# joining all for 2 weeks early
inmet14 = inmet14.withColumnRenamed('UF', 'UF14')
cond = [inmet1.UF == inmet14.UF14, inmet1.epi_week_year == inmet14.epi_week_year_14_early]
inmet_agg = inmet_agg.join(inmet14, cond, 'left')

inmet_agg.filter((F.col('UF')=='11') &  ((F.col('epi_week_year') == '22-2020') | (F.col('epi_week_year') == '23-2020') | (F.col('epi_week_year') == '24-2020') | (F.col('epi_week_year') == '25-2020')))\
    .orderBy('epi_week_year').limit(5).toPandas()

<hr />

# selecting variables of interest
inmet_agg = inmet_agg.select(['UF', 'epi_week_year',
                          'inmet_temp_c_avg', 'inmet_relative_air_humidity_avg', 'inmet_daily_precipt_avg',
                          'inmet_temp_c_1week_before_avg', 'inmet_relative_air_humidity_1week_before_avg', 'inmet_daily_precipt_1week_before_avg',
                          'inmet_temp_c_2weeks_before_avg', 'inmet_relative_air_humidity_2weeks_before_avg', 'inmet_daily_precipt_2weeks_before_avg'])

# padding 'epi_week_year' col
inmet_agg = inmet_agg.withColumn('epi_week_year', F.lpad(F.col('epi_week_year'), 7, '0'))

<hr />

### Joining meteorological data for srags

#### 2021

print("How much distincts before? ", srag_2021.select('NU_NOTIFIC').distinct().count())
print("How much distincts epi_weeks before? ", srag_2021.select('epi_week_year').distinct().count())

srag_2021 = srag_2021.join(inmet_agg, ['UF', 'epi_week_year'], 'left')

print("How much distincts after? ", srag_2021.select('NU_NOTIFIC').distinct().count())
print("How much distincts epi_weeks after? ", srag_2021.select('epi_week_year').distinct().count())

<hr />

#### writing the last temporary version of srags

srag_2021.write.parquet('gs://ai-covid19-datalake/standard/srag/pp_interm_srag_2021_v4_super-srag', mode='overwrite')

<hr />

# # reading temporary files
# srag_2019 = spark.read.parquet('gs://ai-covid19-datalake/standard/srag/pp_interm_srag_2019_v4_super-srag/')
# srag_2020 = spark.read.parquet('gs://ai-covid19-datalake/standard/srag/pp_interm_srag_2020_v4_super-srag')
# srag_2021 = spark.read.parquet('gs://ai-covid19-datalake/standard/srag/pp_interm_srag_2021_v4_super-srag')

srag_2021 = srag_2021.withColumn('AGE_GROUP', F.when(F.col('AGE_AT_NOTIF') < 1, 1)\
                                                    .when((F.col('AGE_AT_NOTIF') >= 1) & (F.col('AGE_AT_NOTIF') <= 5), 2)\
                                                    .when((F.col('AGE_AT_NOTIF') >= 6) & (F.col('AGE_AT_NOTIF') <= 19), 3)\
                                                    .when((F.col('AGE_AT_NOTIF') >= 20) & (F.col('AGE_AT_NOTIF') <= 29), 4)\
                                                    .when((F.col('AGE_AT_NOTIF') >= 30) & (F.col('AGE_AT_NOTIF') <= 39), 5)\
                                                    .when((F.col('AGE_AT_NOTIF') >= 40) & (F.col('AGE_AT_NOTIF') <= 49), 6)\
                                                    .when((F.col('AGE_AT_NOTIF') >= 50) & (F.col('AGE_AT_NOTIF') <= 59), 7)\
                                                    .when((F.col('AGE_AT_NOTIF') >= 60) & (F.col('AGE_AT_NOTIF') <= 69), 8)\
                                                    .when((F.col('AGE_AT_NOTIF') >= 70) & (F.col('AGE_AT_NOTIF') <= 79), 9)\
                                                    .when((F.col('AGE_AT_NOTIF') >= 80) & (F.col('AGE_AT_NOTIF') <= 89), 10)\
                                                    .when((F.col('AGE_AT_NOTIF') >= 90), 11)\
                                                    .otherwise(12))

# just for check the header later
srag_2021.limit(1).toPandas()

# The SUPER SRAG at last

# now the are united, lets create the last attributes
suffix = 'avg'

gmr_inmet_cols = {
    'gmr_transit_stations_': [[-43.09183673469388, -32.61737331954498, -25.20892494929006, -17.24561403508772]],
    'gmr_grocery_and_pharmacy_': [[0.36936936936936937, 8.107558139534884, 13.838709677419354, 19.86090775988287]],
    'gmr_retail_and_recreation_': [[-42.607894736842105, -31.163636363636364, -22.735064935064933, -15.647230320699709]],
    'gmr_workplaces_percent_': [[-15.347786811201445, -7.407114624505929, -4.023725391216558, 0.5605338417540515]],
    'gmr_residential_percent_': [[5.825, 7.780269058295964, 9.963333333333333, 12.902788844621513]],
    'gmr_parks_percent_': [[-45.52965235173824, -36.97651663405088, -30.29383886255924, -21.115384615384617]],
    'gmr_transit_stations_1week_before_': [[-44.09894736842105, -32.61737331954498, -25.263959390862944, -17.85185185185185]],
    'gmr_grocery_and_pharmacy_1week_before_': [[-0.16783216783216784, 7.644859813084112, 13.094644167278062, 18.64406779661017]],
    'gmr_retail_and_recreation_1week_before_': [[-42.81818181818182, -31.208955223880597, -22.889212827988338, -15.928205128205128]],
    'gmr_workplaces_percent_1week_before_': [[-15.347786811201445, -7.351599852887091, -3.9318181818181817, 0.33611111111111114]],
    'gmr_residential_percent_1week_before_': [[5.746001279590531, 7.714285714285714, 9.980842911877394, 13.173153296266879]],
    'gmr_parks_percent_1week_before_': [[-45.42248062015504, -36.97651663405088, -30.672727272727272, -21.639327024185068]],
    'gmr_transit_stations_2weeks_': [[-44.16945606694561, -32.86206896551724, -25.244274809160306, -17.40625]],
    'gmr_grocery_and_pharmacy_2weeks_': [[-0.5689655172413793, 7.173590504451038, 11.925925925925926, 18.068483063328426]],
    'gmr_retail_and_recreation_2weeks_': [[-43.526233359436176, -30.847341337907377, -22.798561151079138, -16.060301507537687]],
    'gmr_workplaces_percent_2weeks_': [[-17.4364406779661, -7.7555555555555555, -4.397341211225997, -0.5694656488549619]],
    'gmr_residential_percent_2weeks_': [[5.692477876106195, 7.639135959339263, 10.018041237113403, 13.193877551020408]],
    'gmr_parks_percent_2weeks_': [[-45.38961038961039, -36.54710144927536, -29.9047131147541, -20.95306859205776]],
    'inmet_temp_c_': [[19.99995412088011, 22.3166005555556, 23.721603579832205, 25.526168289473578]],
    'inmet_relative_air_humidity_': [[65.59553142063525, 69.93693714285736, 73.8520828571426, 78.24404857142858]],
    'inmet_daily_precipt_': [[0.3383720930232559, 1.4355555555555408, 3.35503875968993, 5.962886597938155]],
    'inmet_temp_c_1week_before_': [[20.05500949193533, 22.322388241935233, 23.761091968254064, 25.53099282105264]],
    'inmet_relative_air_humidity_1week_before_': [[65.65009503030261, 70.04159629710138, 73.88706227868933, 78.36411010769187]],
    'inmet_daily_precipt_1week_before_': [[0.3585585585585601, 1.5249999999999975, 3.567741935483844, 6.042424242424034]],
    'inmet_temp_c_2weeks_before_': [[20.174016811110565, 22.320047571428823, 23.821744027322836, 25.50684535398227]],
    'inmet_relative_air_humidity_2weeks_before_': [[65.91730179032191, 70.25798429936523, 73.88706227868934, 78.47414770987979]],
    'inmet_daily_precipt_2weeks_before_': [[0.3714285714285674, 1.6878787878788217, 3.640816326530609, 6.274725274725252]]
}

# Showing the quintiles cuttofs 
for col in pd.DataFrame(gmr_inmet_cols).columns:
    print(col, list(pd.DataFrame(gmr_inmet_cols)[col]))

# Showing the quintiles cuttofs
pd.DataFrame(gmr_inmet_cols)

n_suffix = 'q'
for col in list(gmr_inmet_cols.keys()):
    srag_2021 = srag_2021.withColumn(col+n_suffix, F.when(F.col(col+suffix) <= gmr_inmet_cols[col][0][0], '1')\
                                                                  .when((F.col(col+suffix) > gmr_inmet_cols[col][0][0]) &  (F.col(col+suffix) <= gmr_inmet_cols[col][0][1]), '2')\
                                                                  .when((F.col(col+suffix) > gmr_inmet_cols[col][0][1]) &  (F.col(col+suffix) <= gmr_inmet_cols[col][0][2]), '3')\
                                                                  .when((F.col(col+suffix) > gmr_inmet_cols[col][0][2]) &  (F.col(col+suffix) <= gmr_inmet_cols[col][0][3]), '4')\
                                                                  .when(F.col(col+suffix) > gmr_inmet_cols[col][0][3], '5')\
                                                                  .otherwise('6'))

# making all column names uppercase
for col in srag_2021.columns:
    srag_2021 = srag_2021.withColumnRenamed(col, col.upper())

<hr />
<hr />
<hr />

super_srag = spark.read.parquet('gs://ai-covid19-datalake/standard/super-srag/super_srag_v1.parquet')

super_srag_cols = super_srag.columns

srag_2021 = srag_2021.select(super_srag_cols)

super_srag = super_srag.union(srag_2021)

super_srag = super_srag.withColumn('ANO', F.year('DT_SIN_PRI'))

# srag 2019 has 48554 records
# srag 2020 has 1193735 records
# srag 2021 has 868367 records
# super srag must contain (48554 + 1193735 + 868367 =) 2110656 records
print('super srag has', super_srag.count(), 'records')

#### writing super srag

super_srag.write.parquet('gs://ai-covid19-datalake/standard/super-srag/super_srag_v2.parquet')

super_srag.coalesce(1).write.csv('gs://ai-covid19-datalake/standard/super-srag/super_srag_v2.csv', header=True)
